split app into modules
